Fix DataInitializer.java Linking two documents works as it should

+++ backend/src/main/java/com/kirunaexplorer/app/model/GeoReference.java
@@ -15,7 +15,7 @@ public class GeoReference {
 
     @Id
     @Column(name = "document_id") // Primary key and foreign key
-    private Integer documentId;
+    private Long documentId;
 
     @OneToOne
     @MapsId
